Introduce a tasks module
